Graceful handling of stop pondering
Previously, main search routine was started before pondering was stopped.

This caused a number of issues due to two parallel search routines.

Mainly, expansion was running in parallel on the two sides,
causing some children to be `null`.

We gracefully stop the pondering thread before the main search routine to avoid such issue.